Move WildcardTypeName to TypeName.kt
